backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
rapidfuzz
groq
python-dotenv